+++ pkg/syncer/secret.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/appscode/go/log"
-	"github.com/appscode/kubed/pkg/util"
 	core_util "github.com/appscode/kutil/core/v1"
 	core "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -17,7 +16,7 @@ func (s *ConfigSyncer) SyncSecret(src *core.Secret) error {
 	opt := getSyncOption(src.Annotations)
 
 	if opt.sync { // delete that were in old-ns but not in new-ns and upsert to new-ns
-		newNs, err := util.NamespaceSetForSelector(s.KubeClient, opt.nsSelector)
+		newNs, err := NamespaceSetForSelector(s.KubeClient, opt.nsSelector)
 		if err != nil {
 			return err
 		}
@@ -84,7 +83,7 @@ func (s *ConfigSyncer) syncSecretIntoContexts(src *core.Secret, contexts sets.St
 // upsert into newNs set, delete from (oldNs-newNs) set
 // use skipSrcNs = true for sync in source cluster
 func (s *ConfigSyncer) syncSecretIntoNamespaces(k8sClient kubernetes.Interface, src *core.Secret, newNs sets.String, skipSrcNs bool) error {
-	oldNs, err := util.NamespaceSetForSecretSelector(k8sClient, s.SyncerLabelSelector(src.Name, src.Namespace, s.ClusterName))
+	oldNs, err := NamespaceSetForSecretSelector(k8sClient, s.SyncerLabelSelector(src.Name, src.Namespace, s.ClusterName))
 	if err != nil {
 		return err
 	}
@@ -142,4 +141,18 @@ func (s *ConfigSyncer) upsertSecret(k8sClient kubernetes.Interface, src *core.Se
 	})
 
 	return err
+}
+
+func NamespaceSetForSecretSelector(k8sClient kubernetes.Interface, selector string) (sets.String, error) {
+	secret, err := k8sClient.CoreV1().Secrets(metav1.NamespaceAll).List(metav1.ListOptions{
+		LabelSelector: selector,
+	})
+	if err != nil {
+		return nil, err
+	}
+	ns := sets.NewString()
+	for _, obj := range secret.Items {
+		ns.Insert(obj.Namespace)
+	}
+	return ns, nil
 }
